spec-kit-multi-agent-tdd/requirements.txt
# Requirements for Multi-Agent TDD Spec-Kit

# Template rendering
jinja2>=3.1.0

# Configuration (optional)
pyyaml>=6.0

# Testing
pytest>=9.0.0
